refactor: split graph.py into several components
