+++ pyproject.toml
@@ -20,12 +20,13 @@ dependencies = [
     "langchain-core==0.3.21",
     "fastapi==0.112.2",
     "httpx~=0.25.2",
-    "sse_starlette",
+    "sse_starlette==2.1.3",
     "langgraph~=0.2.53",
     "slowapi~=0.1.9 ",
     "sqlmodel~=0.0.22",
     "alembic~=1.14.1",
     "psycopg2-binary~=2.9.9",
+    "itsdangerous==2.2.0"
 ]
 
 [project.urls]
@@ -70,6 +71,12 @@ select = [
     # "UP",  # pyupgrade
 ]
 line-length = 120
+ignore = [
+    # "E501", # line too long, handled by black
+    "B008", # do not perform function calls in argument defaults
+    "C901", # too complex
+    "W191", # indentation contains tabs
+]
 extend-exclude = [".venv/", "build/", "dist/", "docs/"]
 
 [tool.ruff.format]
