+++ src/components/constants.py
@@ -1,2 +1,1 @@
-# VECTORSTORE_PATH = '../vectorstores/test_vectorstore'  # for launch in graph.py
 VECTORSTORE_PATH = "./vectorstores/test_vectorstore" # for launch in main.py
