+++ src/components/prompts.py
@@ -52,10 +52,10 @@ When recommending supervisors:
    - IF USER SPECIFIED FACULTY, PRESENT HIM ONLY SUPERVISORS FROM SPECIFIED FACULTY
    - Maintain original language of query
    - Never hallucinate - only use retrieved data
-   - WARNING: If you list N supervisors, you MUST explicitly state this number first
-   - IF YOU ARE PLANNING TO PRESENT ONLY N AMOUNT OF SUPERVISORS, DON'T EVEN DARE TO SAY THAT YOU FOUND ANY OTHER AMOUNT OF THEM THAN X
+   - WARNING: If you list N amount of supervisors, you MUST explicitly state this number first
+   - IF YOU ARE PLANNING TO PRESENT ONLY N AMOUNT OF SUPERVISORS, DON'T EVEN DARE TO SAY THAT YOU FOUND ANY OTHER AMOUNT OF THEM THAN N
    - ALWAYS INCLUDE ALL PAPERS AND THESES OF EACH SUPERVISOR IN YOUR OUTPUT
-   - DO NOT REPEAT SUPERVISOR IN YOUR OUTPUT - ONLY UNIQUE SUPERVISOR PER OUTPUT
+   - DO NOT REPEAT SUPERVISOR IN YOUR OUTPUT - ONLY UNIQUE SUPERVISORS PER OUTPUT
 """
 GENERAL_PROMPT_TEMPLATE = """
 Retrieved context: {retrieved_context}
